added validation to course category

+++ app/Http/Controllers/Course/CourseCategoryController.php
@@ -7,6 +7,7 @@ namespace App\Http\Controllers\Course;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\CourseCategoryRequest;
 use App\Models\CourseCategory as CourseCategoryModel;
+use App\Models\Course as CourseModel;
 use App\Utilities\Result;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
@@ -79,6 +80,12 @@ class CourseCategoryController extends Controller
 
         if (! $courseCategory) {
             return $this->result->notFound();
+        }
+
+        // Check if the category to delete have been used to courses
+        $course = CourseModel::where("course_category_id", $id)->firstOrFail();
+        if ($course) {
+            return $this->result->validationError();
         }
 
         $courseCategory->delete();
